28.09.24
Corrected a mistake where unavailable workers would be assigned into the plan because of existing in the paramedics or assistants dictionary after changing the load_backend_function in the previous release where an unavailable worker would only appear and exist in the unavailable attribute and NOwhere else. Now he appears both in either paramedic or assistant attribute even if being unavailable, and also in the unavailable attribute.